:truck: Move test classes to better packaging
Before everything was messy, now it is a little bit more organized
